requirements.txt
# tours.batumi Instagram Automation
# Python 3.10+

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.27.0

# Telegram Bot
python-telegram-bot>=21.0

# Image processing (optional, for photo metadata)
Pillow>=10.0.0

# Video processing (FFmpeg bundled)
imageio-ffmpeg>=0.5.1

# Scheduling
schedule>=1.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
